Extract _zip_update.py from _legacy.py
Moves the Kometa/ImageMaid ZIP-based update primitives and top-level
update flows into modules/helpers/_zip_update.py:

12 small helpers (SHA lookup, download, backup/restore, extraction,
venv, pip install, directory cleanup):
- _ensure_dir, _read_text, _write_text
- _get_upstream_sha, _download_zip
- _backup_kometa_runtime_assets, _restore_kometa_runtime_assets, _cleanup_kometa_backup
- _clear_directory_contents, _extract_zip_bytes
- _ensure_venv, _pip_install

3 top-level update entry points:
- perform_kometa_update_zip_only
- perform_kometa_update_zip_only_at_root
- perform_imagemaid_update_zip_only

Plus the URL constants used by these functions:
- GITHUB_API_BRANCH, GITHUB_ZIP_URL
- IMAGEMAID_GITHUB_API_BRANCH, IMAGEMAID_GITHUB_ZIP_URL

_kometa_version.py updated to import _read_text and _get_upstream_sha
from _zip_update instead of _legacy.

helpers/__init__.py re-exports both public functions (via star import)
and private helpers used by tests (via explicit import). All 1077
tests pass.

Reduces _legacy.py from 1613 to 1107 lines (-506).
